--- a/src/main/java/ch/njol/skript/expressions/ExprColorOf.java
+++ b/src/main/java/ch/njol/skript/expressions/ExprColorOf.java
@@ -19,13 +19,27 @@
  */
 package ch.njol.skript.expressions;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collector;
+import java.util.stream.Collectors;
+
+import org.bukkit.Bukkit;
+import org.bukkit.DyeColor;
+import org.bukkit.FireworkEffect;
+import org.bukkit.Material;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Item;
 import org.bukkit.event.Event;
+import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.material.Colorable;
 import org.bukkit.material.MaterialData;
 import org.eclipse.jdt.annotation.Nullable;
+
+import com.google.common.collect.Lists;
 
 import ch.njol.skript.aliases.ItemType;
 import ch.njol.skript.classes.Changer.ChangeMode;
@@ -34,8 +48,12 @@ import ch.njol.skript.doc.Description;
 import ch.njol.skript.doc.Examples;
 import ch.njol.skript.doc.Name;
 import ch.njol.skript.doc.Since;
-import ch.njol.skript.expressions.base.SimplePropertyExpression;
+import ch.njol.skript.expressions.base.PropertyExpression;
+import ch.njol.skript.lang.Expression;
+import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.util.Color;
+import ch.njol.skript.util.Getter;
+import ch.njol.util.Kleenean;
 import ch.njol.util.coll.CollectionUtils;
 
 /**
@@ -47,41 +65,67 @@ import ch.njol.util.coll.CollectionUtils;
 		"	message \"This wool block is <%colour of block%>%colour of block%<reset>!\"",
 		"	set the colour of the block to black"})
 @Since("1.2")
-public class ExprColorOf extends SimplePropertyExpression<Object, Color> {
+public class ExprColorOf extends PropertyExpression<Object, Color> {
+
 	static {
-		register(ExprColorOf.class, Color.class, "colo[u]r[s]", "itemstacks/entities");
+		register(ExprColorOf.class, Color.class, "colo[u]r[s]", "itemstacks/entities/fireworkeffects");
 	}
 	
 	@SuppressWarnings("null")
 	@Override
-	@Nullable
-	public Color convert(final Object o) {
-		if (o instanceof ItemStack || o instanceof Item) {
-			final ItemStack is = o instanceof ItemStack ? (ItemStack) o : ((Item) o).getItemStack();
-			final MaterialData d = is.getData();
-			if (d instanceof Colorable)
-				return Color.byWoolColor(((Colorable) d).getColor());
-		} else if (o instanceof Colorable) { // Sheep
-			return Color.byWoolColor(((Colorable) o).getColor());
+	public boolean init(Expression<?>[] exprs, int matchedPattern, Kleenean isDelayed, ParseResult parseResult) {
+		setExpr(exprs[0]);
+		return true;
+	}
+	
+	@SuppressWarnings("null")
+	@Override
+	protected Color[] get(Event e, Object[] source) {
+		if (source instanceof FireworkEffect[]) {
+			List<Color> colors = new ArrayList<>();
+			for (FireworkEffect effect : (FireworkEffect[])source) {
+				effect.getColors().stream()
+						.map(color -> Color.byWoolColor(DyeColor.getByColor(color)))
+						.forEach(colour -> colors.add(colour));
+			}
+			if (colors.size() == 0)
+				return null;
+			return colors.toArray(new Color[colors.size()]);
 		}
-		return null;
+		return get(source, new Getter<Color, Object>() {
+			@Override
+			@Nullable
+			public Color get(Object o) {
+				if (o instanceof ItemStack || o instanceof Item) {
+					final ItemStack is = o instanceof ItemStack ? (ItemStack) o : ((Item) o).getItemStack();
+					final MaterialData d = is.getData();
+					if (d instanceof Colorable)
+						return Color.byWoolColor(((Colorable) d).getColor());
+				} else if (o instanceof Colorable) { // Sheep
+					return Color.byWoolColor(((Colorable) o).getColor());
+				}
+				return null;
+			}
+		});
 	}
 	
 	@Override
-	protected String getPropertyName() {
-		return "colour";
-	}
-	
-	@Override
-	public Class<Color> getReturnType() {
+	public Class<? extends Color> getReturnType() {
 		return Color.class;
 	}
 	
-	boolean changeItemStack = false;
+	@Override
+	public String toString(@Nullable Event e, boolean debug) {
+		return "colour of " + getExpr().toString(e, debug);
+	}
+	
+	boolean changeItemStack;
 
 	@Override
 	@Nullable
-	public Class<?>[] acceptChange(final ChangeMode mode) {
+	public Class<?>[] acceptChange(ChangeMode mode) {
+		if (FireworkEffect.class.isAssignableFrom(getExpr().getReturnType()))
+			return CollectionUtils.array(Color[].class);
 		if (mode != ChangeMode.SET)
 			return null;
 		if (Entity.class.isAssignableFrom(getExpr().getReturnType()))
@@ -96,19 +140,15 @@ public class ExprColorOf extends SimplePropertyExpression<Object, Color> {
 	}
 	
 	@Override
-	public void change(final Event e, final @Nullable Object[] delta, final ChangeMode mode) throws UnsupportedOperationException {
-		assert mode == ChangeMode.SET;
-		assert delta != null;
-		
-		final Color c = (Color) delta[0];
-		final Object[] os = getExpr().getArray(e);
-		if (os.length == 0)
+	public void change(Event e, @Nullable Object[] delta, ChangeMode mode) {
+		if (delta == null)
 			return;
 		
-		for (final Object o : os) {
+		Color c = (Color) delta[0];
+		for (Object o : getExpr().getArray(e)) {
 			if (o instanceof ItemStack || o instanceof Item) {
-				final ItemStack is = o instanceof ItemStack ? (ItemStack) o : ((Item) o).getItemStack();
-				final MaterialData d = is.getData();
+				ItemStack is = o instanceof ItemStack ? (ItemStack) o : ((Item) o).getItemStack();
+				MaterialData d = is.getData();
 				if (d instanceof Colorable)
 					((Colorable) d).setColor(c.getWoolColor());
 				else
@@ -124,6 +164,31 @@ public class ExprColorOf extends SimplePropertyExpression<Object, Color> {
 				}
 			} else if (o instanceof Colorable) {
 				((Colorable) o).setColor(c.getWoolColor());
+			} else if (o instanceof FireworkEffect) {
+				Color[] input = (Color[])delta;
+				FireworkEffect effect = ((FireworkEffect) o);
+				switch (mode) {
+					case ADD:
+						for (Color color : input)
+							effect.getColors().add(color.getBukkitColor());
+						break;
+					case REMOVE:
+					case REMOVE_ALL:
+						for (Color color : input)
+							effect.getColors().remove(color.getBukkitColor());
+						break;
+					case DELETE:
+					case RESET:
+						effect.getColors().clear();
+						break;
+					case SET:
+						effect.getColors().clear();
+						for (Color color : input)
+							effect.getColors().add(color.getBukkitColor());
+						break;
+					default:
+						break;
+				}
 			}
 		}
 	}