@TheLimeGlass 2018-10-25
@Blueyescat all your requests are wrong or not adding. There is no point in adding a direction to a firework since a firework can only go upwards, unlike the spawn effect it positions the entity to face a direction.

And placing a space between all the areas you want will break as Skript doesn't allow spaces in it's classinfo names!