@Blueyescat 2018-10-25
> There is no point in adding a direction to a firework since a firework can only go upwards, unlike the spawn effect it positions the entity to face a direction.
No, the spawned entity follows yaw and pitch of the location. You can do ```vb
spawn cow right player``` or something like this when it is `[%directions% %locations%]`.

> And placing a space between all the areas you want will break as Skript doesn't allow spaces in it's classinfo names!

Alrady said that they are syntaxes not *classinfo name*s.