@TheBentoBox 2018-10-25
> a firework can only go upwards

This is untrue and is becoming even more untrue in survival gameplay come 1.14 since [crossbows can shoot fireworks in any direction](https://youtu.be/6y6vdqfKPYQ?t=902).